5.4.21-20260918_17897218

+++ polardbx-test/src/test/java/com/alibaba/polardbx/qatest/ddl/cdc/CdcImplicitTableGroupTest.java
@@ -148,6 +148,10 @@ public class CdcImplicitTableGroupTest extends CdcBaseTest {
 
     @SneakyThrows
     private void executeWithCallBack(Consumer<Parameter> consumer, String dbName, String createDbSql) {
+        if (!supportImplicitTableGroup()) {
+            return;
+        }
+
         try (Statement stmt = tddlConnection.createStatement()) {
             stmt.executeUpdate("drop database if exists " + dbName);
             stmt.executeUpdate(createDbSql);
